ExprOperator now type checks
