+++ src/ast/expr/mod.rs
@@ -1,5 +1,7 @@
 mod expr;
 mod literal;
+mod operator;
 
 pub use expr::*;
 pub use literal::*;
+pub use operator::*;
